+++ packages/@eventual/core-runtime/src/clients/workflow-client.ts
@@ -12,23 +12,23 @@ import {
   WorkflowExecutionOptions,
 } from "@eventual/core";
 import {
-  hashCode,
-  StartExecutionResponse,
   WorkflowEventType,
-  WorkflowStarted,
+  type StartExecutionResponse,
+  type WorkflowStarted,
 } from "@eventual/core/internal";
 import { ulid } from "ulidx";
 import { inspect } from "util";
-import {
-  formatExecutionId,
-  INTERNAL_EXECUTION_ID_PREFIX,
-} from "../workflow/execution.js";
 import { WorkflowSpecProvider } from "../providers/workflow-provider.js";
 import { computeScheduleDate } from "../schedule.js";
 import { ExecutionStore } from "../stores/execution-store.js";
 import { createEvent } from "../workflow/events.js";
+import {
+  formatExecutionId,
+  INTERNAL_EXECUTION_ID_PREFIX,
+} from "../workflow/execution.js";
 import { ExecutionQueueClient } from "./execution-queue-client.js";
 import { LogsClient } from "./logs-client.js";
+import { hashCode } from "../utils.js";
 
 export class WorkflowClient {
   constructor(
