+++ packages/@eventual/core-runtime/src/index.ts
@@ -3,7 +3,6 @@ export * from "./call-executor.js";
 export * from "./clients/index.js";
 export * from "./console-hook.js";
 export * from "./eventual-hook.js";
-export * from "./workflow/execution.js";
 export * from "./handlers/index.js";
 export * from "./local/index.js";
 export * from "./log-agent.js";
@@ -19,4 +18,5 @@ export * from "./task-token.js";
 export * from "./tasks.js";
 export * from "./transaction-executor.js";
 export * from "./utils.js";
+export * from "./workflow/execution.js";
 export * from "./workflow/index.js";
