+++ packages/@eventual/core-runtime/src/utils.ts
@@ -1,10 +1,11 @@
-import type {
+import {
   Attributes,
   BucketNotificationEvent,
   Entity,
   EntityCompositeKeyPart,
   EntityStreamItem,
   KeyTuple,
+  SystemError,
 } from "@eventual/core";
 import type {
   BucketNotificationHandlerSpec,
@@ -158,4 +159,43 @@ export async function streamToBuffer(stream: Readable) {
   }
 
   return Buffer.concat(chunks);
+}
+
+/**
+ * Returns a hash code from a string
+ * @param  {String} str The string to hash.
+ * @return {Number}    A 32bit integer
+ * @see http://werxltd.com/wp/2010/05/13/javascript-implementation-of-javas-string-hashcode-method/
+ * @see https://stackoverflow.com/a/8831937/968011
+ */
+export function hashCode(str: string): number {
+  let hash = 0;
+  for (let i = 0, len = str.length; i < len; i++) {
+    const chr = str.charCodeAt(i);
+    hash = (hash << 5) - hash + chr;
+    hash |= 0; // Convert to 32bit integer
+  }
+  return hash;
+}
+
+export function extendsError(err: unknown): err is Error {
+  return (
+    !!err &&
+    typeof err === "object" &&
+    (err instanceof Error ||
+      ("prototype" in err &&
+        !!err.prototype &&
+        Object.prototype.isPrototypeOf.call(err.prototype, Error)))
+  );
+}
+
+export function extendsSystemError(err: unknown): err is SystemError {
+  return (
+    !!err &&
+    typeof err === "object" &&
+    (err instanceof SystemError ||
+      ("prototype" in err &&
+        !!err.prototype &&
+        Object.prototype.isPrototypeOf.call(err.prototype, SystemError)))
+  );
 }
