+++ packages/@eventual/core-runtime/src/workflow/call-eventual-factory.ts
@@ -4,6 +4,7 @@ import {
   type Call,
   type CallOutput,
 } from "@eventual/core/internal";
+import { Result } from "../result.js";
 import { AwaitTimerClassEventualFactory } from "./call-executors-and-factories/await-timer-call.js";
 import { BucketCallEventualFactory } from "./call-executors-and-factories/bucket-store-call.js";
 import { ChildWorkflowCallEventualFactory } from "./call-executors-and-factories/child-workflow-call.js";
@@ -12,18 +13,28 @@ import { EmitEventsCallEventualFactory } from "./call-executors-and-factories/em
 import { EntityCallEventualFactory } from "./call-executors-and-factories/entity-store-call.js";
 import { ExpectSignalFactory } from "./call-executors-and-factories/expect-signal-call.js";
 import { SearchCallEventualFactory } from "./call-executors-and-factories/open-search-client-call.js";
-import { RegisterSignalHandlerCallFactory } from "./call-executors-and-factories/register-signal-handler-call.js";
 import { SendSignalEventualFactory } from "./call-executors-and-factories/send-signal-call.js";
+import { RegisterSignalHandlerCallFactory } from "./call-executors-and-factories/signal-handler-call.js";
 import { TaskCallEventualFactory } from "./call-executors-and-factories/task-call.js";
 import { TransactionCallEventualFactory } from "./call-executors-and-factories/transaction-call.js";
 import { UnsupportedEventualFactory } from "./call-executors-and-factories/unsupported.js";
 import type { EventualDefinition } from "./eventual-definition.js";
 
+export interface ResolveEventualFunction {
+  (seq: number, result: Result): void;
+}
+
 /**
  * Turns a {@link Call} into an {@link EventualDefinition} for a {@link Workflow}.
  */
 export interface EventualFactory<C extends Call = Call> {
-  createEventualDefinition(call: C): EventualDefinition<CallOutput<C>>;
+  initializeEventual(
+    call: C,
+    /**
+     * Call this function to resolve another eventual with a value.
+     */
+    resolveEventual: ResolveEventualFunction
+  ): EventualDefinition<CallOutput<C>>;
 }
 
 export type AllWorkflowEventualFactories = {
@@ -34,14 +45,17 @@ export type AllWorkflowEventualFactories = {
 
 export class AllWorkflowEventualFactory implements EventualFactory {
   constructor(private executors: AllWorkflowEventualFactories) {}
-  public createEventualDefinition(call: Call): EventualDefinition<any> {
+  public initializeEventual(
+    call: Call,
+    resolveEventual: ResolveEventualFunction
+  ): EventualDefinition<any> {
     const kind = call[CallSymbol];
     const executor = this.executors[CallKind[kind] as keyof typeof CallKind] as
       | EventualFactory
       | undefined;
 
     if (executor) {
-      return executor.createEventualDefinition(call);
+      return executor.initializeEventual(call, resolveEventual);
     }
 
     throw new Error(`Missing Eventual Factory for ${CallKind[kind]}`);
@@ -61,7 +75,7 @@ export function createDefaultEventualFactory(): AllWorkflowEventualFactory {
     ExpectSignalCall: new ExpectSignalFactory(),
     GetExecutionCall: unsupportedFactory,
     InvokeTransactionCall: new TransactionCallEventualFactory(),
-    RegisterSignalHandlerCall: new RegisterSignalHandlerCallFactory(),
+    SignalHandlerCall: new RegisterSignalHandlerCallFactory(),
     SearchCall: new SearchCallEventualFactory(),
     SendSignalCall: new SendSignalEventualFactory(),
     StartWorkflowCall: unsupportedFactory,
