+++ packages/@eventual/core-runtime/src/workflow/call-executors-and-factories/emit-events-call.ts
@@ -1,10 +1,10 @@
 import {
-  Result,
   WorkflowCallHistoryType,
   type EmitEventsCall,
 } from "@eventual/core/internal";
 import type { EventualFactory } from "../call-eventual-factory.js";
 import type { EventualDefinition } from "../eventual-definition.js";
+import { Result } from "../../result.js";
 
 /**
  * Create a event for the {@link EmitEventsCall} and return undefined.
@@ -14,9 +14,7 @@ import type { EventualDefinition } from "../eventual-definition.js";
 export class EmitEventsCallEventualFactory
   implements EventualFactory<EmitEventsCall>
 {
-  public createEventualDefinition(
-    call: EmitEventsCall
-  ): EventualDefinition<void> {
+  public initializeEventual(call: EmitEventsCall): EventualDefinition<void> {
     return {
       createCallEvent: (seq) => ({
         type: WorkflowCallHistoryType.EventsEmitted,
