+++ packages/@eventual/core/src/internal/index.ts
@@ -1,13 +1,11 @@
-export * from "./bucket-hook.js";
 export * from "./calls.js";
 export * from "./entity.js";
 export * from "./eventual-hook.js";
 export * from "./eventual-service.js";
-export * from "./global.js";
 export * from "./guards.js";
 export * from "./open-api-spec.js";
 export * from "./properties.js";
-export * from "./result.js";
+export * from "./resources.js";
 export * from "./schedule.js";
 export * from "./service-spec.js";
 export * from "./service-type.js";
